requirements.txt
Flask==3.0.3
line-bot-sdk==3.11.0
python-dotenv==1.0.1
gunicorn==22.0.0
